backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
numpy==1.26.2
trimesh==4.0.8
rembg==2.0.56
git+https://github.com/VAST-AI-Research/TripoSR.git
